Move TestDevice to Core to reuse it in TestNG and JUnit
